pyproject.toml
[project]
name = "pdf-summarizer"
version = "0.1.0"
description = "pdf processor & summarizer"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.20",
    "langchain-anthropic>=0.3.10",
    "langchain-community>=0.3.19",
    "langchain-mistralai>=0.2.8",
    "pypdf>=5.3.1",
    "python-dotenv>=1.0.1",
    "transformers>=4.49.0",
]
